feat: basic egui implementation
Added basic egui implementation using eframe.
