+++ Cargo.toml
@@ -6,6 +6,9 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+eframe = "0.18.0"
+egui = "0.18.1"
+egui_extras = { version = "0.18.0", features = ["image"] }
 #gtk4 = "0.4.7"
 image = "0.24.2"
 #libadwaita = "0.1.1"
